Minor: fixed returning of long of big numbers (instead of negatives one)

+++ core/src/main/java/com/orientechnologies/orient/core/index/hashindex/local/cache/O2QCache.java
@@ -804,7 +804,7 @@ public class O2QCache implements OReadCache {
 
   @Override
   public long getUsedMemory() {
-    return (am.size() + a1in.size()) * (2 * ODurablePage.PAGE_PADDING + pageSize);
+    return ((long) (am.size() + a1in.size())) * (2 * ODurablePage.PAGE_PADDING + pageSize);
   }
 
   private OCacheEntry remove(long fileId, long pageIndex) {
